backend/requirements.txt
altair==5.5.0
altex==0.2.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.59
botocore==1.42.59
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
deap==1.4.3
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
entrypoints==0.4
et_xmlfile==2.0.0
Faker==40.5.1
fastapi==0.109.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
favicon==0.7.0
filelock==3.25.0
fonttools==4.61.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==0.6.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lightgbm==4.3.0
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
markdownlit==0.0.7
MarkupSafe==3.0.3
matplotlib==3.8.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
narwhals==2.17.0
numpy==1.26.4
nvidia-nccl-cu12==2.29.7
openai==2.24.0
openpyxl==3.1.2
packaging==24.2
pandas==2.2.0
pillow==11.3.0
platformdirs==4.9.2
plotly==6.5.2
postgrest==2.28.2
prometheus_client==0.24.1
propcache==0.4.1
protobuf==5.29.6
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.11.0
pymdown-extensions==10.21
pyOpenSSL==25.3.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.3
realtime==2.28.2
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
scikit-learn==1.4.1.post1
scipy==1.11.3
seaborn==0.12.2
sentry-sdk==2.54.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==4.3.0
snowflake-snowpark-python==1.45.0
sortedcontainers==2.4.0
soupsieve==2.8.3
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.36.3
stopit==1.1.2
storage3==2.28.2
streamlit==1.42.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-shadcn-ui==0.1.19
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.4
TPOT==0.11.7
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.6.3
uvicorn==0.27.1
uvloop==0.22.1
validators==0.35.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==14.2
wheel==0.46.3
xgboost==2.0.3
yarl==1.23.0
zstandard==0.25.0
